✨ feat: UserRepository 통합 및 회원가입/로그인/이메일 인증 API 연동

+++ data/src/main/java/com/example/data/api/dto/server/JoinDTO.kt
@@ -17,7 +17,7 @@ data class JoinDTO (
     val gender: Int,
 
     @Json(name = "jobId")
-    val jobId: Long,
+    val jobId: Int,
 
     @Json(name = "purposeList")
     val purposeList: List<String>? = null,
